@@ -1,24 +1,26 @@
-var React = require('react');
-var classNames = require('classnames');
-var blacklist = require('blacklist');
+const React = require('react');
+const classNames = require('classnames');
+const blacklist = require('blacklist');
+const PropTypes = require('prop-types');
+const createReactClass = require('create-react-class');
 
-const InputGroup = React.createClass({
+const InputGroup = createReactClass({
 	displayName: 'InputGroup',
 	propTypes: {
-		className: React.PropTypes.string,
-		contiguous: React.PropTypes.bool
+		className: PropTypes.string,
+		contiguous: PropTypes.bool,
 	},
-	render() {
+	render () {
 		// classes
-		var className = classNames('InputGroup', {
-			'InputGroup--contiguous': this.props.contiguous
+		const className = classNames('InputGroup', {
+			'InputGroup--contiguous': this.props.contiguous,
 		}, this.props.className);
-		var props = blacklist(this.props, 'contiguous');
+		const props = blacklist(this.props, 'contiguous');
 
 		return (
 			<div {...props} className={className} />
 		);
-	}
+	},
 });
 
 // expose the child to the top level export
