@@ -1,14 +1,16 @@
-var classnames = require('classnames');
-var React = require('react');
+const classnames = require('classnames');
+const React = require('react');
+const PropTypes = require('prop-types');
+const createReactClass = require('create-react-class');
 
-export default React.createClass({
+export default createReactClass({
 	displayName: 'ModalFooter',
 	propTypes: {
-		children: React.PropTypes.node.isRequired,
-		className: React.PropTypes.string
+		children: PropTypes.node.isRequired,
+		className: PropTypes.string,
 	},
-	render() {
+	render () {
 		var className = classnames('Modal__footer', this.props.className);
 		return <div {...this.props} className={className} />;
-	}
+	},
 });
