@@ -1,15 +1,17 @@
-var classnames = require('classnames');
-var React = require('react');
-var blacklist = require('blacklist');
+const classnames = require('classnames');
+const React = require('react');
+const blacklist = require('blacklist');
+const PropTypes = require('prop-types');
+const createReactClass = require('create-react-class');
 
-export default React.createClass({
+export default createReactClass({
 	displayName: 'ModalHeader',
 	propTypes: {
-		children: React.PropTypes.node,
-		className: React.PropTypes.string,
-		onClose: React.PropTypes.func,
-		showCloseButton: React.PropTypes.bool,
-		text: React.PropTypes.string,
+		children: PropTypes.node,
+		className: PropTypes.string,
+		onClose: PropTypes.func,
+		showCloseButton: PropTypes.bool,
+		text: PropTypes.string,
 	},
 	handleClose () {
 		document.body.style.overflow = '';
@@ -18,10 +20,10 @@ export default React.createClass({
 	render () {
 
 		// elements
-		var className = classnames('Modal__header', this.props.className);
-		var close = this.props.showCloseButton ? <button type="button" onClick={this.handleClose} className="Modal__header__close" /> : null;
-		var text = this.props.text ? <h4 className="Modal__header__text">{this.props.text}</h4> : null;
-		var props = blacklist(this.props, 'children', 'onClose', 'showCloseButton', 'text');
+		const className = classnames('Modal__header', this.props.className);
+		const close = this.props.showCloseButton ? <button type="button" onClick={this.handleClose} className="Modal__header__close" /> : null;
+		const text = this.props.text ? <h4 className="Modal__header__text">{this.props.text}</h4> : null;
+		const props = blacklist(this.props, 'children', 'onClose', 'showCloseButton', 'text');
 
 		return (
 			<div {...props} className={className}>
