@@ -1,61 +1,63 @@
-var React = require('react');
-var blacklist = require('blacklist');
-var classNames = require('classnames');
+const React = require('react');
+const blacklist = require('blacklist');
+const classNames = require('classnames');
+const PropTypes = require('prop-types');
+const createReactClass = require('create-react-class');
 
-export default React.createClass({
+export default createReactClass({
 	displayName: 'RadioGroup',
 	propTypes: {
-		alwaysValidate: React.PropTypes.bool,
-		className: React.PropTypes.string,
-		inline: React.PropTypes.bool,
-		label: React.PropTypes.string,
-		onChange: React.PropTypes.func.isRequired,
-		options: React.PropTypes.array.isRequired,
-		required: React.PropTypes.bool,
-		requiredMessage: React.PropTypes.string,
-		value: React.PropTypes.string
+		alwaysValidate: PropTypes.bool,
+		className: PropTypes.string,
+		inline: PropTypes.bool,
+		label: PropTypes.string,
+		onChange: PropTypes.func.isRequired,
+		options: PropTypes.array.isRequired,
+		required: PropTypes.bool,
+		requiredMessage: PropTypes.string,
+		value: PropTypes.string,
 	},
-	getDefaultProps() {
+	getDefaultProps () {
 		return {
-			requiredMessage: 'This field is required'
+			requiredMessage: 'This field is required',
 		};
 	},
-	getInitialState() {
+	getInitialState () {
 		return {
 			isValid: true,
-			validationIsActive: this.props.alwaysValidate
+			validationIsActive: this.props.alwaysValidate,
 		};
 	},
-	componentDidMount() {
+	componentDidMount () {
 		if (this.state.validationIsActive) {
 			this.validateInput(this.props.value);
 		}
 	},
-	componentWillReceiveProps(newProps) {
+	componentWillReceiveProps (newProps) {
 		if (this.state.validationIsActive) {
 			if (newProps.value !== this.props.value && newProps.value !== this._lastChangeValue && !newProps.alwaysValidate) {
 				// reset validation state if the value was changed outside the component
 				return this.setState({
 					isValid: true,
-					validationIsActive: false
+					validationIsActive: false,
 				});
 			}
 			this.validateInput(newProps.value);
 		}
 	},
-	handleChange(e) {
+	handleChange (e) {
 		this._lastChangeValue = e.target.value;
 		if (this.props.onChange) this.props.onChange(e.target.value);
 	},
-	handleBlur() {
+	handleBlur () {
 		if (!this.props.alwaysValidate) {
 			this.setState({ validationIsActive: false });
 		}
 		this.validateInput(this.props.value);
 	},
-	validateInput(value) {
+	validateInput (value) {
 		var newState = {
-			isValid: true
+			isValid: true,
 		};
 		if (this.props.required && (!value || (value && !value.length))) {
 			newState.isValid = false;
@@ -65,7 +67,7 @@ export default React.createClass({
 		}
 		this.setState(newState);
 	},
-	render() {
+	render () {
 		var self = this;
 
 		// props
@@ -73,7 +75,7 @@ export default React.createClass({
 
 		// classes
 		var componentClass = classNames('FormField', {
-			'is-invalid': !this.state.isValid
+			'is-invalid': !this.state.isValid,
 		}, this.props.className);
 
 		// validation message
@@ -88,7 +90,7 @@ export default React.createClass({
 		var componentLabel = this.props.label ? <label className="FormLabel">{this.props.label}</label> : null;
 
 		// options
-		var radios = this.props.options.map(function(radio, i) {
+		var radios = this.props.options.map(function (radio, i) {
 			return (
 				<label key={'radio-' + i} className="Radio">
 					<input value={radio.value} type="radio" onChange={self.handleChange} onBlur={self.handleBlur} name={self.props.name} className="Radio__input" />
@@ -107,5 +109,5 @@ export default React.createClass({
 				{validationMessage}
 			</div>
 		);
-	}
+	},
 });
